Disable migration of long inserters
## What
Removed the migration of long inserters into fast inserters found in this migration JSON file.

## Why
The reason being that it messes with blueprints too much, and there are no upgraded versions of the long inserters specifically, with the default longer pickup and insert positions.

If this migration is left in, it will "break" a lot of blueprints, as there is no swap-in replacement for the long inserter, due to the reasons mentioned above.

### Break?
What I mean by "break" is; blueprints will no longer function as expected, as the pickup and insert positions of the mk3 inserter (the long inserter's replacement) is different to the long inserter.

## Future solutions
- Add a startup setting to enable the replacing of the long inserters for those that want it.
- Create upgraded tiers of specifically the long inserters.